feat: implement useCalculateWaitingTime hook and refactor OrderManagement and TableStatusManagement to utilize it
